+++ NEWS.md
@@ -38,6 +38,8 @@
 * `mapply` was reimplemented using variadic templates. `mapply` now accepts a 
   function as first parameter, then a variable number of sugar expressions. 
 
+* `Array` gains a `fill` method to initialize all its data to the same value.    
+  
 # Rcpp11 3.1.0.1
 
 * `Rcpp11` is now compatible with the gcc 4.6.3 as shipped by Rtools. Therefore